Update job function

+++ api/src/main/java/com/bluecollar/jobmanagement/repo/JobRepository.java
@@ -3,6 +3,6 @@ package com.bluecollar.jobmanagement.repo;
 import org.springframework.data.repository.CrudRepository;
 import com.bluecollar.jobmanagement.model.Job;
 
-public interface JobRepository extends CrudRepository<Job, Long>
+public interface JobRepository extends CrudRepository<Job, Integer>
 {
 }
